chore(model): fix pre-existing clippy lints (Wave 1.6)
11 trivial lint cleanups in vllm-model test files and lib tests:
- 5x empty_line_after_outer_attr in qwen3_token_pipeline.rs
- unused import ModelBackend in qwen3_integration.rs
- unused Module in qwen3_5/model_tests.rs
- approx_constant 0.693_147 -> std::f32::consts::LN_2 in ssm.rs
- identity_op '+ 0' in gemma4/attention.rs
- unnecessary_cast 'as usize' in qwen3_5/speculative_tests.rs
  (also adds #[allow(clippy::needless_range_loop)] on the enclosing
   test fn, since clippy 1.94 fires the stricter form when step is
   used as a direct index without a cast — equivalent semantics)
- items_after_test_module: move impl Default above mod tests in gemma4/attention.rs

All pre-existing (introduced before Wave 1); no behavior change.
Unblocks `cargo clippy --workspace --all-targets -- -D warnings`.

+++ crates/model/src/qwen3_5/speculative_tests.rs
@@ -162,6 +162,7 @@ fn test_forward_to_layer_matches_full_forward_full_attention_only() {
 }
 
 #[test]
+#[allow(clippy::needless_range_loop)]
 fn test_gdn_prefill_decode_sequence_is_stable() {
     let config = tiny_config(vec!["linear_attention", "linear_attention"]);
     let device = Device::Cpu;
@@ -192,7 +193,7 @@ fn test_gdn_prefill_decode_sequence_is_stable() {
             *ref_chain.last().unwrap(),
             num_computed,
         );
-        assert_eq!(next, decoded[step as usize]);
+        assert_eq!(next, decoded[step]);
         ref_chain.push(next);
     }
 }
